backend: ajustado para retornar todas as igrejas do usuario no momento do login

+++ backend/app/Http/Controllers/Api/LoginController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
+use App\Http\Resources\ChurchResource;
 use App\Models\User;
 use Exception;
 use Illuminate\Http\Request;
@@ -11,16 +12,30 @@ use Illuminate\Support\Facades\Auth;
 class LoginController extends Controller
 {
     public function login(Request $request) {
-        if (Auth::attempt(['email' => $request->email, 'password' => $request->password])) {
+        $credentials = $request->only('email', 'password');
 
+        if (Auth::attempt($credentials)) {
             $user = Auth::user();
+
+            if ($user->person) {
+                $members = $user->person->member()->with('churches')->get();
+
+                $churches = $members->flatMap(function ($member) {
+                    return $member->churches;
+                })->unique('id');
+
+                $churchesArray = $churches->toArray();
+            } else {
+                $churchesArray = [];
+            }
 
             $token = $request->user()->createToken('token')->plainTextToken;
 
             return response()->json([
                 'status' => true,
                 'token' => $token,
-                'user' => $user
+                'user' => $user,
+                'churches' => $churchesArray,
             ], 201);
         } else {
             return response()->json([
@@ -33,14 +48,15 @@ class LoginController extends Controller
     public function logout(User $user) {
         try {
 
-            $user->tokens()->delete();
+            if ($user->tokens()->count() > 0) {
+                $user->tokens()->delete();
+            }
 
             return response()->json([
                 'status' => true,
                 'message' => 'Deslogado'
             ], 200);
         } catch (Exception $e) {
-
             return response()->json([
                 'status' => false,
                 'message' => 'Não deslogado'
